% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAccessTokenByUsernameAndPassword.R
\name{getAccessTokenByUsernameAndPassword}
\alias{getAccessTokenByUsernameAndPassword}
\title{Creates an API-Access-Token on Gitlab.Pavlovia.Org by using username/password.}
\usage{
getAccessTokenByUsernameAndPassword(username, password)
}
\arguments{
\item{username}{Username or Email of a registered account in Pavlovia}

\item{password}{Passord for the Pavlovia/Gitlab.Pavlovia Account}
}
\value{
A Response object with $data (Names list: $name-> Name of Token , $token -> AccessToken) and $message a status message and $isError (True if API Call was unsuccessful)
}
\description{
Tokens named PavloviaR_#Random_String#. Tokens are saved to accessToken.csv in the current working directory and will be reused.
So we avoid creating an unlimited number of accessTokens on gitlab.pavlovia.org. Be careful with the token!
}
\examples{
tokenResponseObject <- getAccessTokenByUsernameAndPassword("username", "12345")
tokenResponseObject$data$token #here is your accessToken
}
