% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTibbleOfDataDirectoryOfProject.R
\name{getTibbleOfDataDirectoryOfProject}
\alias{getTibbleOfDataDirectoryOfProject}
\title{Title}
\usage{
getTibbleOfDataDirectoryOfProject(accessToken, projectID)
}
\arguments{
\item{projectID}{}
}
\value{

}
\description{
Title
}
