% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeDataTibble.R
\name{mergeDataTibble}
\alias{mergeDataTibble}
\title{Merging all available Data-Files with more than 0 rows (with same and different column sizes)}
\usage{
mergeDataTibble(data)
}
\arguments{
\item{data}{List of Data-Files in Data-Directory generated by getTibbleOfDataDirectoryOfProject}
}
\value{
Merged Data-Files as a long tibble
}
\description{
You have to filter the list of data-tibbles to merge a subset of the files
}
\examples{
mergedData <- mergeDataTibble(getTibbleOfDataDirectoryOfProject("#######",12345))
}
