% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAccessTokenByUsernameAndPassword.R
\name{getAccessTokenByUsernameAndPassword}
\alias{getAccessTokenByUsernameAndPassword}
\title{Title}
\usage{
getAccessTokenByUsernameAndPassword(username, password)
}
\arguments{
\item{password}{}
}
\value{

}
\description{
Title
}
